requirements.txt
django==2.2.9
djangorestframework>=3.9.0,<3.10.0
django-filter
Markdown
psycopg2-binary
Pillow
gunicorn
django-debug-toolbar
django-extensions
django-pdb
docutils
Jinja2
jinja2-django-tags
MarkupSafe
pytz
raven
django-taggit
social-auth-app-django
sorl-thumbnail
flake8>=3.6.0,<3.7.0
